pyproject.toml
[build-system]
requires = ["setuptools==65.6.3"]
build-backend = "setuptools.build_meta"

[project]
name = "nerfview"
version = "0.0.0"
authors = []
requires-python = ">=3.8"
dependencies = [
  "viser @ git+https://github.com/nerfstudio-project/viser@main",
  "jaxtyping",
]

[project.optional-dependencies]
examples = [
  "tyro",
  "nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@main",
  "roma",
  "gsplat @ git+https://github.com/nerfstudio-project/gsplat.git@v1.0",
  "imageio[ffmpeg]",
  "scikit-learn",
  "torchmetrics[image]",
  "opencv-python",
  "tensorboard",
  "pycolmap @ git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e",
]
dev = ["black", "isort", "ipdb"]

[tool.setuptools.packages.find]
include = ["nerfview"]

[tool.isort]
profile = "black"
